refactor: reorganize imports and move addPullRequestLabel to init directory
