+++ src/CLI.ts
@@ -3,7 +3,7 @@
 import { Command, Option } from "commander";
 import { checkNpmTestCompletion } from "./checkNpmTestCompletion.js";
 import { cleanMerged } from "./cleanMerged.js";
-import { addReleaseNoteTemplate } from "./init/addReleaseNoteTemplate.js";
+import { addReleaseNoteTemplate } from "./init/index.js";
 import { initLabel } from "./initLabel.js";
 import { postversion } from "./postversion.js";
 import { preversion } from "./preversion.js";
