+++ src/init/index.ts
@@ -1,0 +1,2 @@
+export { addReleaseNoteTemplate } from "./addReleaseNoteTemplate.js";
+export { addPullRequestLabel } from "./addPullRequestLabel.js";
