+++ src/initLabel.ts
@@ -1,4 +1,4 @@
-import { addPullRequestLabel } from "./addPullRequestLabel.js";
+import { addPullRequestLabel } from "./init/index.js";
 
 export async function initLabel() {
 	await addPullRequestLabel(
